+++ packages/credentials/vitest.config.ts
@@ -1,6 +1,10 @@
 import { defineConfig } from "vitest/config";
+
+import { coverageConfig } from "../../vitest.shared";
+
 export default defineConfig({
   test: {
     globals: true,
+    coverage: coverageConfig,
   },
 });